Rutas de diary protegidas con JWT

+++ src/modules/diary/diary.service.ts
@@ -3,6 +3,7 @@ import { Diary } from "./diary.model";
 import { DiaryRepository } from "./diary.repository"
 import { createDiaryInput } from "./diary.schema"
 import { NotFoundError } from "../../shared/errors/notFound";
+import { AppError } from "../../shared/errors/app.error";
 
 export class DiaryService{
     private readonly diaryRepository: DiaryRepository;
@@ -12,9 +13,10 @@ export class DiaryService{
         this.diaryRepository = diaryRepository;
     }
 
-    async create(input: createDiaryInput): Promise<string|null>{
+    async create(input: createDiaryInput, creatorId: string): Promise<string|null>{
         const diary: Diary = {
             id: randomUUID(),
+            creatorId: creatorId,
             content: input.content,
             createdAt: new Date(),
             updatedAt: new Date()
@@ -39,11 +41,17 @@ export class DiaryService{
 
     }
 
-    async update(id:string, content: string): Promise<void>{
+    //find by user id
+
+    async update(id:string, content: string, userId: string): Promise<void>{
         const existing = await this.diaryRepository.findById(id);
 
         if(!existing){
             throw new NotFoundError("Diary");
+        }
+
+        if(existing.creatorId !== userId){
+            throw new AppError("You dont have permission", 401, "UNAUTHORIZED");
         }
 
         const updated: Diary = {
@@ -56,11 +64,15 @@ export class DiaryService{
 
     }
 
-    async deleteById(id: string): Promise<void>{
+    async deleteById(id: string, userId: string): Promise<void>{
         const existing = await this.diaryRepository.findById(id);
 
         if(!existing){
             throw new NotFoundError("Diary");
+        }
+
+        if(existing.creatorId !== userId){
+            throw new AppError("You dont have permission", 401, "UNAUTHORIZED");
         }
 
         return this.diaryRepository.deleteById(id);
